Split `supa.db` into `supa.db.model` and `supa.db.session`.
Only `db_session` needs to be imported locally (due to how application
initialization works). No sense in requiring the same for all DB models
just because they happen to be defined in the same module. By splitting
them up into separate modules this restricting is lifted.
